% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{classify_notes_to_experiment_taster}
\alias{classify_notes_to_experiment_taster}
\title{Classify Notes According to Experiment (Tastermonial Edition)}
\usage{
classify_notes_to_experiment_taster(foodname)
}
\arguments{
\item{foodname}{character string name of the food item}
}
\value{
character vector with canonical name for the food item
}
\description{
Same results as \code{\link[=classify_notes_to_experiment]{classify_notes_to_experiment()}} but uses pre-canned Tastermonial mapping
}
