% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{db_delete_user_from_table}
\alias{db_delete_user_from_table}
\title{Delete Rows For Specific \code{user_id}}
\usage{
db_delete_user_from_table(con, tablename, user_id)
}
\arguments{
\item{con}{valid database connection}

\item{user_id}{user ID}
}
\description{
Delete Rows For Specific \code{user_id}
}
