% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{db_update_records}
\alias{db_update_records}
\title{Update All Records in \code{table_name} that match \code{user_id}}
\usage{
db_update_records(con, user_id, table_name, table_df)
}
\arguments{
\item{con}{valid database connection}

\item{user_id}{user ID}

\item{table_name}{character string name of a database table}

\item{table_df}{dataframe of records to substitute}
}
\value{
integer number of rows affected
}
\description{
Update All Records in \code{table_name} that match \code{user_id}
}
