% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create.R
\name{deploy_add_table}
\alias{deploy_add_table}
\title{Deployment : Add a New Table}
\usage{
deploy_add_table(proposed_table = "accounts_user")
}
\arguments{
\item{proposed_table}{character string indicated one of the \code{data()} objects available to this package.}
}
\description{
To be used only interactively at deployment time, this function will add a new table to the
currently connected database (as determined by \code{db_connection()}). Please be super careful to name
the \code{proposed_table} to be something that already exists as a \code{data()} object in this package.
}
