% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{experiments_mapping}
\alias{experiments_mapping}
\title{Experiments Mapping}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 93 rows and 2 columns.
}
\usage{
experiments_mapping
}
\description{
Users might write food names in slightly different ways, so to
make data cleanup easier, packages can use the \code{experiments_mapping} table
to recognize patterns.
}
\keyword{datasets}
