% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create.R
\name{make_new_database_if_necessary}
\alias{make_new_database_if_necessary}
\title{Create or delete a new database as required}
\usage{
make_new_database_if_necessary(
  conn_args = config::get("dataconnection"),
  db_name,
  drop = FALSE,
  force = FALSE
)
}
\arguments{
\item{conn_args}{connection}

\item{db_name}{character string name for proposed new database}

\item{drop}{Nuke a database with this name if it already exists (default = \code{FALSE})}

\item{force}{override the warning about creating a database with the magic name \code{qsdb} (default = \code{FALSE})}
}
\description{
Intended for situations where you want to start all over,
this function will use direct SQL calls to either create a brand new database
or, if optionally \code{drop=TRUE}, wipe out any existing database.
IMPORTANT: database connection is dropped regardless, so if you need a connection,
be sure reconnect.
}
