% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create.R
\name{cgm_db}
\alias{cgm_db}
\title{Initialize a database object}
\usage{
cgm_db(db_config = "sqldb")
}
\arguments{
\item{db_config}{string name of configuration to be used  for this instance.}
}
\description{
With this object, there is no need to keep track of database connections in order to access
the Tastermonial database. Once initialized, you simply use the \code{$} methods to pull the glucose and notes
records results. If the database doesn't exist, initializing this object will create it for you.
}
\section{Functions}{
\itemize{
\item \code{cgm_db}: returns a valid table
}}

