% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{write_notes_raw_from_scratch}
\alias{write_notes_raw_from_scratch}
\title{Drop Notes and Write From Scratch}
\usage{
write_notes_raw_from_scratch(con, notes_df, dry_run = FALSE)
}
\arguments{
\item{con}{valid database connection}

\item{notes_df}{table to write to notes records}

\item{dry_run}{logical: if true, then don't make any permanent changes}
}
\description{
Drop Notes and Write From Scratch
}
